sambar: audio_amplifier: correct the stop_threshold used for calibration
Prevent triggering an automatic stop of the stream when the
buffer empties (ie. is not replenished fast enough by the
write_dummy_data thread) by setting the stop_threshold to be
arbitrarily large. This is similar to what is already done in
the reference calibration code.

This avoids the kernel warning that the stream has already been
stopped when the /dev/snd/pcm* node being used for calibration
is closed:

 WARNING: at ./sound/soc/msm/qdsp6v2/msm-pcm-q6-v2.c:479 msm_pcm_trigger+0xfc/0x180()

SAMBAR-970
Change-Id: Icd92a3c4e4ffa12665859b4e2158fb0c3f81416d