sambar: fix line continuation error in makefile generation
setup-makefiles.sh would throw an extra line continuation character
at the end of PRODUCT_COPY_FILES if proprietary-files.txt ended with
a file that should not be included in PRODUCT_COPY_FILES.  This
caused the final endif to be included in PRODUCT_COPY_FILES, rather
than ending the makefile if clause.  Fix it by calculating the number
of lines needing a trailing '\' as

(total lines) - (lines starting with [-#] or blank)

When generating the PRODUCT_COPY_FILES, only iterate through lines
that aren't ignored (starting with [-#] or blank).

Change-Id: Ieaf5e3aa2ab0b29c01383b9474d819d8fc2edcb2