keyhandler: fix torch gesture sometimes toggle twice
handleKeyEvent function pass KeyEvent structure directly
to the event handler, but actually from log that structure
will be reused by caller, so passing that structure to other
message handler is unsafe and mess up the ACTION_UP/DOWN in
handleMessage.

Another problem is in Proximity on Wake, it stops process
new message when there is already key message pending to mEventHandler.
Making sometimes only the down event is fired to mEventHandler.

SAMBAR-1096

Change-Id: Idb3a1135bb1484a954a07fb0a0909438c14ae5b7