sambar: keyhandler: perform haptic feedback before target operation
A goal of haptic feedback is to quickly tell the user when a gesture
was recognized. The previous code performed the vibration after performing
the action.  In the case of flashlight, this could be 100s of ms later.
Instead, perform haptic feedback first to inform the user as quickly
as possible that an action will occur.

Change-Id: I84fa4fc00363c2e8e529737dd40f6af4b13ee98f